upgrade_pythoncapi: try to preserve newlines when patching
In case we use \n on Windows or \r\n on Linux we don't want
patching those files change every line due to newlines being
adjusted to the platform defaults.

Instead pass 'newline=""' to all open() calls to preserve
newlines. And when adding new lines use the first type of
newline found in the file.

The test changes make the patching code reuseable and adds
a second test with multiple different newlines in the input.